added old code of the table form for reference and stylizing, see the comment block in the landlord-page-tables. this may never see the light of day.  but for more information this is a cross post to what I sent to elton about thw landlordpage and the landlord reviews table
one thing, the old code on <landlord-page.js> was more for the list of all individual reviews themselves instead of the overall.  I think im going to put the old code back perhaps in another file or commented out to work as a building block for that list of reviews.  It doesnt have to be a table by any means, but just a metric for showing individual reviews themselves.

but the end result will be the same in regards to, displaying a landlord page that has a chunk of their overall (what you did) stats, then underneath (either by loading in a component or otherwise) display the amalgamation of individual reviews.

gratatata